+++ src/EventHub/EventHub.Test/ScenarioTests/Namespace.cs
@@ -32,13 +32,6 @@ namespace Microsoft.Azure.Commands.EventHub.Test.ScenarioTests
         }
 
         [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void NamespaceAuthorizationRulesCRUD()
-        {
-            TestRunner.RunTestScript("NamespaceAuthTests");
-        }
-
-        [Fact]
         [Trait(Category.AcceptanceType, Category.LiveOnly)]
         public void MSIEncryptionTest()
         {
@@ -50,20 +43,6 @@ namespace Microsoft.Azure.Commands.EventHub.Test.ScenarioTests
         public void MSITest()
         {
             TestRunner.RunTestScript("MSITest");
-        }
-
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void SchemaRegistry()
-        {
-            TestRunner.RunTestScript("SchemaRegistryTest");
-        }
-
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void ApplicationGroupTest()
-        {
-            TestRunner.RunTestScript("ApplicationGroupTest");
         }
     }
 }